add functional testing for the new redirects
These tests demonstrate that an item like Course, which
has no shipping method, will not redirect the user to
shipping-method. Rather the user will be taken directly
to the payment method url.
The opposite test is also included.
The fixtures include just one product with it's own category,
a Course, which has no shipping method